+++ src/lib/notificationsRequests.js
@@ -7,10 +7,11 @@ const notificationUrl = 'notification';
 const usersUrl        = 'users';
 
 // --- ONE-LINE REQUEST FUNCTIONS ---
-export const createNotification = (data)       => API.post(notificationUrl, data);
-export const userNotifications  = (id)         => API.get(`${notificationUrl}/${id}`);
-export const removeNotification = (id)         => API.delete(`${notificationUrl}/${id}`);
-export const updateNotification = (id, data)   => API.patch(`${notificationUrl}/${id}`, data);
+export const createNotification = (data, token)       => API.post(notificationUrl, data, { headers: { Authorization: `Bearer ${token}` } });
+export const userNotifications  = (id, token, query = '') => API.get(`${notificationUrl}/${id}${query}`, { headers: { Authorization: `Bearer ${token}` } });
+export const removeNotification = (id, token)         => API.delete(`${notificationUrl}/${id}`, { headers: { Authorization: `Bearer ${token}` } });
+export const updateNotification = (id, data, token)   => API.patch(`${notificationUrl}/${id}`, data, { headers: { Authorization: `Bearer ${token}` } });
+export const readAllNotificationsReq = (headers) => API.patch(`${notificationUrl}/read-all`, {}, { headers });
 
 export const getUser = (userId, token) => {
     const headers = {
